Move transaction verifier to its own crate
