+++ chainstate/src/detail/block_index_history_iter.rs
@@ -13,8 +13,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+use chainstate_types::BlockIndexHandle;
 use common::{chain::GenBlock, primitives::Id};
-use consensus::BlockIndexHandle;
 use logging::log;
 
 use super::GenBlockIndex;
