+++ chainstate/src/detail/chainstateref/mod.rs
@@ -13,12 +13,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use std::{collections::BTreeSet, convert::TryInto, sync::Arc};
+use std::{collections::BTreeSet, convert::TryInto};
 
 use chainstate_storage::{BlockchainStorageRead, BlockchainStorageWrite, TransactionRw};
 use chainstate_types::{
-    get_skip_height, storage_result, BlockIndex, GenBlockIndex, GetAncestorError,
-    PropertyQueryError,
+    block_index_ancestor_getter, get_skip_height, BlockIndex, BlockIndexHandle, GenBlockIndex,
+    GetAncestorError, PropertyQueryError,
 };
 use common::{
     chain::{
@@ -33,12 +33,14 @@ use common::{
     time_getter::TimeGetterFn,
     Uint256,
 };
-use consensus::{BlockIndexHandle, TransactionIndexHandle};
+use consensus::TransactionIndexHandle;
 use logging::log;
 use utils::{ensure, tap_error_log::LogError};
 use utxo::{UtxosDB, UtxosStorageRead, UtxosStorageWrite, UtxosView};
 
 use crate::{BlockError, BlockSource, ChainstateConfig};
+
+use self::tx_verifier_storage::gen_block_index_getter;
 
 use super::{
     median_time::calculate_median_time_past,
@@ -1001,76 +1003,5 @@ impl<'a, S: BlockchainStorageWrite, O: OrphanBlocksMut> ChainstateRef<'a, S, O> 
             Ok(_) => Ok(()),
             Err(err) => err.into(),
         }
-    }
-}
-
-pub fn block_index_ancestor_getter<S, G>(
-    gen_block_index_getter: G,
-    db_tx: &S,
-    chain_config: &ChainConfig,
-    block_index: &GenBlockIndex,
-    target_height: BlockHeight,
-) -> Result<GenBlockIndex, GetAncestorError>
-where
-    G: Fn(&S, &ChainConfig, &Id<GenBlock>) -> Result<Option<GenBlockIndex>, storage_result::Error>,
-{
-    if target_height > block_index.block_height() {
-        return Err(GetAncestorError::InvalidAncestorHeight {
-            block_height: block_index.block_height(),
-            ancestor_height: target_height,
-        });
-    }
-
-    let mut height_walk = block_index.block_height();
-    let mut block_index_walk = block_index.clone();
-    loop {
-        assert!(height_walk >= target_height, "Skipped too much");
-        if height_walk == target_height {
-            break Ok(block_index_walk);
-        }
-        let cur_block_index = match block_index_walk {
-            GenBlockIndex::Genesis(_) => break Ok(block_index_walk),
-            GenBlockIndex::Block(idx) => idx,
-        };
-
-        let ancestor = cur_block_index.some_ancestor();
-
-        let height_walk_prev =
-            height_walk.prev_height().expect("Can never fail because prev is zero at worst");
-        let height_skip = get_skip_height(height_walk);
-        let height_skip_prev = get_skip_height(height_walk_prev);
-
-        // prepare the booleans for the check
-        let at_target = height_skip == target_height;
-        let still_not_there = height_skip > target_height;
-        let too_close = height_skip_prev.next_height().next_height() < height_skip;
-        let prev_too_close = height_skip_prev >= target_height;
-
-        if at_target || (still_not_there && !(too_close && prev_too_close)) {
-            block_index_walk = gen_block_index_getter(db_tx, chain_config, ancestor)
-                .log_err()?
-                .expect("Block index of ancestor must exist, since id exists");
-            height_walk = height_skip;
-        } else {
-            let prev_block_id = cur_block_index.prev_block_id();
-            block_index_walk = gen_block_index_getter(db_tx, chain_config, prev_block_id)
-                .log_err()?
-                .ok_or(GetAncestorError::PrevBlockIndexNotFound(*prev_block_id))
-                .log_err()?;
-            height_walk = height_walk_prev;
-        }
-    }
-}
-
-pub fn gen_block_index_getter<S: BlockchainStorageRead>(
-    db_tx: &S,
-    chain_config: &ChainConfig,
-    block_id: &Id<GenBlock>,
-) -> Result<Option<GenBlockIndex>, storage_result::Error> {
-    match block_id.classify(chain_config) {
-        GenBlockId::Genesis(_id) => Ok(Some(GenBlockIndex::Genesis(Arc::clone(
-            chain_config.genesis_block(),
-        )))),
-        GenBlockId::Block(id) => db_tx.get_block_index(&id).map(|b| b.map(GenBlockIndex::Block)),
     }
 }
