+++ chainstate/tx-verifier/src/transaction_verifier/amounts_map.rs
