+++ chainstate/tx-verifier/src/transaction_verifier/cached_operation.rs
