+++ chainstate/tx-verifier/src/transaction_verifier/mod.rs
@@ -34,7 +34,7 @@ use fallible_iterator::FallibleIterator;
 
 use std::collections::{btree_map::Entry, BTreeMap};
 
-use chainstate_types::{BlockIndex, GenBlockIndex};
+use chainstate_types::{block_index_ancestor_getter, BlockIndex, GenBlockIndex};
 use common::{
     amount_sum,
     chain::{
@@ -54,8 +54,6 @@ use self::token_issuance_cache::TokenIssuanceCache;
 
 mod utils;
 use self::utils::{check_transferred_amount, get_input_token_id_and_amount};
-
-use super::chainstateref::block_index_ancestor_getter;
 
 // TODO: We can move it to mod common, because in chain config we have `token_min_issuance_fee`
 //       that essentially belongs to this type, but return Amount
