+++ chainstate/types/Cargo.toml
@@ -10,6 +10,6 @@ license = "MIT"
 serialization = { path = "../../serialization" }
 common = { path = "../../common/" }
 storage = { path = "../../storage/" }
-thiserror = "1.0"
 
+thiserror = "1.0"
 parity-scale-codec = "3.1"
