+++ chainstate/types/src/block_index_handle.rs
@@ -13,7 +13,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use chainstate_types::{BlockIndex, GenBlockIndex, PropertyQueryError};
+use crate::{BlockIndex, GenBlockIndex, PropertyQueryError};
 use common::{
     chain::{block::BlockReward, Block, GenBlock},
     primitives::{BlockHeight, Id},
